requirements.txt
aiohttp==3.9.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
tldextract==5.1.1
pandas==2.1.4
networkx==3.2.1
rapidfuzz==3.5.2
extruct==0.13.0
w3lib==2.1.1
jsonschema==4.20.0
tenacity==8.2.3
aiolimiter==1.1.0
playwright==1.40.0
urllib3==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0


